cycle edit

+++ ws.h
@@ -1,3 +1,3 @@
 #include "config.h"
 int compute_cycles_gemm_ws(config* config, result *result);
-int compute_util_gemm_ws(config* config, result *result);
+float compute_util_gemm_ws(config* config, result *result);
